backend/requirements.txt
flask
flask-cors
yfinance
python-dotenv

# Full version (Google OAuth + token exchange) — uncomment when re-enabling OAuth:
# requests
